refactor(directory): classify feature about game
