+++ docs/README.md
@@ -61,3 +61,6 @@
 - 시도할 횟수 입력과 게임 결과 `container` 보이기
   - 불필요할 수 있지만 기능의 명확성을 위해 `container` 보이기 기능을 묶어 분리하였다.
 
+- 디렉토리 구조 - 모듈
+  - `car/actions.js` 에 구현한 자동차 이동 기록 기능과 최종 우승자 결정 기능은 자동차에 관련된 기능이 아니라고 생각되어서 다시 모듈화 하였다.
+  - 게임에 관련된 기능을 모아서 `game/actions.js` 을 새로 만들었다.
