+++ src/index.js
@@ -1,10 +1,5 @@
 import { getCarNames, getRacingCount } from './user/actions.js';
-import {
-  setRacingCars,
-  moveRacingCars,
-  checkRacingLog,
-  pickWinner,
-} from './car/actions.js';
+import play from './game/actions.js';
 
 const carNamesSubmitBtn = document.getElementById('car-names-submit');
 const racingCountSubmitBtn = document.getElementById('racing-count-submit');
@@ -15,20 +10,6 @@ const gameResult = document.getElementById('game-result');
 export default function RacingCarGame() {
   let carNames;
   let racingCount;
-
-  const play = function (_carNames, _racingCount) {
-    const racingCars = setRacingCars(_carNames);
-    let racingResult = '';
-
-    for (let i = 0; i < _racingCount; i++) {
-      moveRacingCars(racingCars);
-      racingResult += checkRacingLog(racingCars);
-    }
-
-    racingResult += pickWinner(racingCars);
-
-    return racingResult;
-  };
 
   const showHiddenContainer = function (_container) {
     _container.hidden = false;
